src/main/java/ru/netology/domain/Radio.java
package ru.netology.domain;

public class Radio {

    private int currentNumberStation;
    private int maxNumberStation = 9;
    private int minNumberStation = 0;

    public void increaseCurrentNumberStation() {
        if (currentNumberStation >= maxNumberStation) {
            currentNumberStation = minNumberStation;
        }
        currentNumberStation++;
    }

    public void decreaseCurrentNumberStation() {
        if (currentNumberStation <= minNumberStation) {
            currentNumberStation = maxNumberStation;
        }
        currentNumberStation--;
    }

    public int getCurrentNumberStation() {
        return currentNumberStation;
    }

    public void setCurrentNumberStation(int currentNumberStation) {
        this.currentNumberStation = currentNumberStation;
    }
}
